backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2024.8.30
click==8.1.7
databases==0.9.0
django-extensions==3.2.3
fastapi==0.115.6
h11==0.14.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
sqlparse==0.5.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.33.0
